Move FileUtils to libs
