+++ src/libs/localFileCreate/index.native.ts
@@ -1,5 +1,5 @@
 import RNFetchBlob from 'react-native-blob-util';
-import * as FileUtils from '@libs/fileDownload/FileUtils';
+import * as FileUtils from '@libs/FileUtils';
 import type LocalFileCreate from './types';
 
 /**
